MetaStation GATO Overhaul V3
This time I prioritized the atmospherics pipes concerning engineering, the AI SAT, the rest of science, and some of the southern halls. As well as touching up some, but not all, areas of science including Toxins/Ordinance. I aim to touch the main R&D lab, the RD's office, and robotics in a future patch.